24 authentication not working

+++ app/routers/news_items.py
@@ -47,7 +47,7 @@ async def get_newsitem(id: int, r: Request, database: Database):
     return newsitem
 
 @router.post("/", response_model=NewsItemResponse)
-async def create_newsitem(data: NewsItemCreate, database: Database, active_user: Annotated[User, Depends(current_user)]):
+async def create_newsitem(data: NewsItemCreate, database: Database, active_user: Annotated[User, Depends(get_current_user)]):
     t = now()
     try:
         newsitem = NewsItem.model_validate(data, update={'created_at': t, 'updated_at': t, 'user_id': active_user.id })
@@ -106,11 +106,10 @@ async def get(id: int, database: Database):
     return comments.all()
 
 @router.post("/{id}/comments", response_model=NewsCommentResponse)
-async def create_comment(data: NewsCommentCreate, database: Database, active_user: Annotated[User, Depends(current_user)]):
+async def create_comment(data: NewsCommentCreate, database: Database, active_user: Annotated[User, Depends(get_current_user)]):
     t = now()
     try:
-        comment = NewsComment.model_validate(data, update={'created_at': t, 'updated_at': t, 'user_id': 313 })
-        # comment = NewsComment.model_validate(data, update={'created_at': t, 'updated_at': t, 'user_id': active_user.id })
+        comment = NewsComment.model_validate(data, update={'created_at': t, 'updated_at': t, 'user_id': active_user.id })
     except ValidationError as error:
         raise HTTPException(status_code=500, detail="Input data not valid")
     database.add(comment)
@@ -121,15 +120,15 @@ async def create_comment(data: NewsCommentCreate, database: Database, active_use
 
 
 @router.delete("/comments/{comment_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_comment(comment_id: int, database: Database, active_user: Annotated[User, Depends(current_user)]):
+async def delete_comment(comment_id: int, database: Database, active_user: Annotated[User, Depends(get_current_user)]):
     comment:NewsComment | None = await database.get(NewsComment, comment_id)
 
     if not comment:
         raise HTTPException(status_code=404, detail="NewsComment not found")
     # For now only allow the owner of the comment to remove it
     # TODO: Admins and owner of comment can remove comments
-    # if comment.user_id != active_user.id:
-    #     raise HTTPException(status_code=403, detail="Not authorized to delete this comment")
+    if comment.user_id != active_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to delete this comment")
     await database.delete(comment)
     await database.commit()
     return
